Remove dead code.
I might well be very wrong here, so please review carefully. Removing this code
does not change anything to the tests results, so it is definitely either dead
code or several test cases are missing. For example, could someone explain what
"support patterns" means in the context of rest parameters? I would be glad to
get some insights on how to write tests that this commit make fail.